add monkey patch and rename
